pyproject.toml
[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devscript"
version = "0.1.0"
description = "A utility for managing custom development scripts via JSON or TOML configuration"
requires-python = ">=3.8"
license = {text = "GPL-3.0-only"}

readme = "README.md"
authors = [
    {name = "Маг Ильяс DOMA (MagIlyasDOMA)", email = "magilyas.doma.09@list.ru"}
]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]

keywords = ["cli", "scripts"]
dependencies = [
    "pyyaml>=6.0.3",
]

[project.urls]
Homepage = "https://github.com/MagIlyasDOMA/devscript"
Repository = "https://github.com/MagIlyasDOMA/devscript.git"
BugTracker = "https://github.com/MagIlyasDOMA/devscript/issues"

[project.scripts]
devscript = "devscript:main"
devscr = "devscript:main"
devs = "devscript:main"
dvs = "devscript:main"

[tool.setuptools]
packages = ["devscript"]
